+++ src/config/celery.py
@@ -4,6 +4,6 @@ from celery import Celery
 
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "config.settings")
 
-app = Celery("docker_django")
+app = Celery("doc_django")
 app.config_from_object("django.conf:settings", namespace="CELERY")
 app.autodiscover_tasks()