+++ src/config/urls.py
@@ -1,5 +1,5 @@
 """
-URL configuration for docker_django project.
+URL configuration for doc_django project.
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/4.2/topics/http/urls/